+++ main.py
@@ -6,14 +6,12 @@ from fastapi import FastAPI, Depends, HTTPException, status
 from typing import List
 
 from src.db.connect import get_db
-from src.models.entity import Base, User, Message
-from src.repositories.users import UserRepository
-from src.repositories.messages import MessageRepository
-from src.schemas.schemas import UserCreate, UserLogin, MessageCreate, MessageResponse
-from src.routes import referrals
+from src.routes import referrals, users
+
 
 app = FastAPI(swagger_ui_parameters={"operationsSorter": "method"})
 app.include_router(referrals.router, prefix="/api", tags=["referrals"])
+app.include_router(users.router, prefix="/api", tags=["users"])
 
 @app.get("/", name="root")
 def read_root():
@@ -48,34 +46,6 @@ def healthchecker(db: Session = Depends(get_db)):
         print(e)
         raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
                             detail="Error connecting to the database")
-    
-
-@app.post("/users/", response_model=UserCreate)
-def create_user(user: UserCreate, db: Session = Depends(get_db)):
-    user_repo = UserRepository(db)
-    db_user = user_repo.get_by_telegram_id(user.telegram_id)
-    if db_user:
-        raise HTTPException(status_code=400, detail="User already registered")
-    
-    new_user = User(
-        username=user.username,
-        email=user.email,
-        password=user.password,
-        telegram_id=user.telegram_id,
-        first_name=user.first_name,
-        language_code=user.language_code
-    )
-
-    user_repo.add(new_user)
-    return new_user
-
-@app.post("/login/")
-def login_user(user: UserLogin, db: Session = Depends(get_db)):
-    user_repo = UserRepository(db)
-    if user_repo.validate_user(user.email, user.password):
-        return {"message": "Login successful"}
-    else:
-        raise HTTPException(status_code=400, detail="Invalid credentials")
 
 if __name__ == '__main__':
     uvicorn.run(app="main:app", reload=True, host="127.0.0.1", port=8000)
