+++ src/schemas/users.py
@@ -3,20 +3,12 @@ from typing import Optional
 
 class UserCreate(BaseModel):
     username: str
-    email: str
     password: str
-    telegram_id: int
+    tg_id: int
     first_name: str
     language_code: str
 
 class UserLogin(BaseModel):
-    email: str
     password: str
 
-class MessageCreate(BaseModel):
-    message_id: int
-    user_id: int
-    chat_id: int
-    text: str
-    command: Optional[str] = None
